when typing enter and inserting a paragraph, don't accidentally delete an inserted empty paragraph

+++ source/tbird/chrome/content/bidimailpack/bidimailpack-composer.js
@@ -344,9 +344,12 @@ function InsertParagraph()
   var par = findClosestBlockElement(editor.selection.focusNode);
   var prevPar = par.previousSibling;
 
+  alert('prevpar = ' + prevPar.innerHTML);
+  alert('par = ' + par.innerHTML);
+
   // Hunt and shoot the extra BR. We don't want it.
   var node = prevPar.lastChild;
-  if (node && (node.nodeType == node.ELEMENT_NODE) && (node.tagName.toUpperCase() == "BR"))
+  if (node && (node.nodeType == node.ELEMENT_NODE) && (node.tagName.toUpperCase() == "BR") && (prevPar.childNodes.length > 1))
     editor.deleteNode(node);
 
   editor.endTransaction();
